backend/requirements.txt
(rezwan) C:\Users\rezwa\OneDrive\Desktop\DevConnect\DevConnect\backend>pip list
Package                       Version
----------------------------- -------
asgiref                       3.9.1
Django                        5.2.5
django-cors-headers           4.7.0
djangorestframework           3.16.1
djangorestframework_simplejwt 5.5.1
mysqlclient                   2.2.7
pillow                        11.3.0
pip                           24.0
PyJWT                         2.10.1
PyMySQL                       1.1.2
sqlparse                      0.5.3
tzdata                        2025.2